[BUG] queue: don't dequeue proxy-global requests on disabled servers
If a server is disabled or tracking a disabled server, it must not
dequeue requests pending in the proxy queue, it must only dequeue
its own ones.

The problem that was caused is that if a backend always had requests
in its queue, a disabled server would continue to take traffic forever.
(cherry picked from commit 09d02aaf02d1f21c0c02672888f3a36a14bdd299)