[MEDIUM] servers: support address 0.0.0.0 as the original destination address
Until now, if a server had no address (or 0.0.0.0), it would have tried
to connect to that address, which results in a connection to localhost
on some systems, sometimes leading to a loop. In theory it should not
have been possible to do that but a check was missing.

Now what will happen is that haproxy will try to connect to the same IP
as the one the original connection was sent to. This means that it will
do the exact same thing as the "transparent" mode, except that a server
also supports connection limiting and health checks, which are not possible
with "transparent".

It also allows to force the port but not the IP, which was not possible
in the case of the transparent method.
(cherry picked from commit 1b393a9c302637419bfaedcba6bc67b6cb44cdf0)