[MINOR] config: fix endianness of server check port
The "addr" server setting supports an undocumented optional port
part which is used if the "port" setting is not set. However, it
is incorrectly used as it's not correctly converted from big endian,
so eventhough it's not documented, let's fix it.
(cherry picked from commit 5400d112489a3e27218976b44dece940980b93b6)