[MINOR] http: keep pointer to beginning of data
We now set msg->col and msg->sov to the first byte of non-header.
They will be used later when parsing chunks. A new macro was added
to perform size additions on an http_msg in order to limit the risks
of copy-paste in the long term.

During this operation, it appeared that the http_msg struct was not
optimal on 64-bit, so it was re-ordered to fill the holes.
(cherry picked from commit fa355d4a51072ef2f0bc37147d7048e6b758cab0)

Note for 1.3: this is not a bugfix but this change is required to
get the upcoming cookie bugfixes.