[MEDIUM] http: fix space handling in the response cookie parser
This patch addresses exactly the same issues as the previous one, but
for responses this time. It also introduces implicit support for the
Set-Cookie2 header, for which there's almost nothing specific to do
since it is a clean header. This one allows multiple cookies in a
same header, by respecting the HTTP messaging semantics.

The new parser has been tested with insertion, rewrite, passive,
removal, prefixing and captures, and it looks OK. It's still able
to rewrite (or delete) multiple cookies at once. Just as with the
request parser, it tries hard to fix formating of the cookies it
displaces.

This patch too should be backported to 1.4 and possibly to 1.3.
(cherry picked from commit 24581bae022bcf97ea7818e49ef27d21c92d6aa3)
(cherry picked from commit 4f54636fccaefd961cc0c537462260a3242e3648)

Important note for 1.3:
   This version of the patch includes the fix from commit bd59bef7,
   because the original patch introduces a regression. So this version
   of the patch is safe.